+++ pmm/proxysql_metrics.go
@@ -58,7 +58,7 @@ func (a *Admin) AddProxySQLMetrics(dsn string) error {
 	srv := consul.AgentService{
 		ID:      serviceID,
 		Service: "proxysql:metrics",
-		Tags:    []string{fmt.Sprintf("alias_%s", a.ServiceName)},
+		Tags:    []string{fmt.Sprintf("alias_%s", a.ServiceName), "scheme_https"},
 		Port:    int(port),
 	}
 	reg := consul.CatalogRegistration{
@@ -75,9 +75,16 @@ func (a *Admin) AddProxySQLMetrics(dsn string) error {
 		Value: []byte(SanitizeDSN(dsn))}
 	a.consulAPI.KV().Put(d, nil)
 
+	// Check and generate certificate if needed.
+	if err := a.checkSSLCertificate(); err != nil {
+		return err
+	}
+
 	args := []string{
 		fmt.Sprintf("-web.listen-address=%s:%d", a.Config.BindAddress, port),
 		fmt.Sprintf("-web.auth-file=%s", ConfigFile),
+		fmt.Sprintf("-web.ssl-cert-file=%s", SSLCertFile),
+		fmt.Sprintf("-web.ssl-key-file=%s", SSLKeyFile),
 	}
 
 	// Install and start service via platform service manager.